+++ src/topology/cpu/cache.rs
@@ -1,6 +1,6 @@
 //! CPU cache hierarchy discovery.
 
-#[cfg(feature = "std")]
+#[cfg(all(feature = "std", any(windows, target_os = "linux")))]
 use super::CacheLevel;
 
 #[cfg(all(feature = "std", target_os = "linux"))]
@@ -8,7 +8,7 @@ mod linux;
 #[cfg(all(feature = "std", windows))]
 mod windows;
 
-#[cfg(feature = "std")]
+#[cfg(all(feature = "std", any(windows, target_os = "linux")))]
 pub(crate) fn detect_cache_levels(_logical_processors: usize) -> Option<Box<[CacheLevel]>> {
     #[cfg(target_os = "linux")]
     {
@@ -18,10 +18,5 @@ pub(crate) fn detect_cache_levels(_logical_processors: usize) -> Option<Box<[Cac
     #[cfg(windows)]
     {
         windows::detect()
-    }
-
-    #[cfg(not(any(target_os = "linux", windows)))]
-    {
-        None
     }
 }